+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## master
+## 1.11.0 (2020-05-17)
 
 * [minimagick] Handle destination format having no file extension (@janko)